fix(ai): ItemRewrite scores are float, not int
Sonnet/Haiku occasionally emit decimal scores (e.g. 7.5) despite the
prompt asking for 0-10. With int|None, Pydantic raises ValidationError
on any fractional response — losing an already-billed Anthropic result
and forcing a re-pay on the next pass.

float|None accepts both 7 and 7.5 (Pydantic auto-coerces ints). The
existing v2 pipeline (results.py) already uses float for the same
fields, so this also removes a v2/v3 inconsistency. Downstream JSONB
casts via `::int` round numerics correctly, so the calibration query
in spec §3.5 is unaffected.

Add three tests pinning the new contract: 7.5 accepted, 2.5 accepted,
integers still accepted.

+++ src/docket/ai/rewrite_schema.py
@@ -15,9 +15,9 @@ class ItemRewrite(BaseModel):
     headline: str | None = Field(None, max_length=60)
     why_it_matters: str | None = Field(None, max_length=200)
     significance_rationale: str = Field("", max_length=1500)
-    significance_score: int | None = Field(None, ge=0, le=10)
+    significance_score: float | None = Field(None, ge=0.0, le=10.0)
     consent_placement_rationale: str = Field("", max_length=1500)
-    consent_placement_score: int | None = Field(None, ge=0, le=10)
+    consent_placement_score: float | None = Field(None, ge=0.0, le=10.0)
     suggested_badge_slugs: list[str] = Field(default_factory=list)
     confidence: Literal['high', 'medium', 'low']
